accomodate multiple days

+++ workflow/extract-photo-metadata.R
@@ -13,7 +13,9 @@ pblapply(flight_days, FUN = function(x) {
   photo_list <- list.files(x, full.names = TRUE, recursive = TRUE)
   flight_day_basename <- basename(x)
   
-  metadata <- exifr::read_exif(path = photo_list)
-
-  write.csv(x = metadata, file = paste0("data/drone/L0/metadata/", flight_day_basename, "_metadata.csv"), row.names = FALSE)
-  })
+  if(!file.exists(paste0("data/drone/L0/metadata/", flight_day_basename, "_metadata.csv"))) {
+    metadata <- exifr::read_exif(path = photo_list)
+    
+    write.csv(x = metadata, file = paste0("data/drone/L0/metadata/", flight_day_basename, "_metadata.csv"), row.names = FALSE)
+  }
+})
